requirements.txt
streamlit>=1.36
google-generativeai>=0.7.2
python-dotenv>=1.0.1
PyYAML>=6.0.2
